feat(middleware): refactor middleware for easier testing
- Introduced a Middleware interface for better abstraction
- Updated routes to accept middleware via dependency injection
- Use real middleware in main application
- Use fake/test middleware in integration tests
